backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
qrcode==7.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
